Define assert using ... so that the argument can contain a comma (C23).
This affects cases where the argument expression to assert contains an un-parenthesized comma, which can occur when using compound literals.

When NDEBUG is not defined, this still requires the argument to be something that would count as one argument to a function (i.e. it cannot use an un-parenthesized comma operator). We do this by including an unevaluated call to a fake function using that argument as part of the expansion of the assert macro. (This behavior is specified by a note in the standard, and it's not entirely clear to me that it follows from the normative text, but I've implemented it that way since that's apparently the intent.)